splited main.js into modules
